backend/requirements.txt
# -------------------------
# Core API Framework
# -------------------------
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
email-validator==2.1.1

# -------------------------
# Data Processing & ML
# -------------------------
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
joblib==1.4.2
SQLAlchemy==2.0.32
python-dateutil==2.9.0.post0

# -------------------------
# Deep Learning & Imaging
# -------------------------
pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
medmnist==3.0.2

# -------------------------
# File Uploads & Parsing
# -------------------------
python-multipart==0.0.9
pypdf2==3.0.1

# -------------------------
# OCR Support
# -------------------------
pytesseract==0.3.10
pdf2image==1.17.0
regex==2024.5.15
pytz==2024.1

# -------------------------
# PDF Export / Reporting
# -------------------------
reportlab==4.2.0
pycairo==1.26.0

# -------------------------
# CLI Tools & Utilities
# -------------------------
tqdm==4.66.5

# -------------------------
# Authentication / Security
# -------------------------
bcrypt==4.1.2
python-jose==3.3.0
python-dotenv==1.0.1
passlib==1.7.4
